Split up impls for MAC and DMA, and require that SMI has exclusive access to the entirety of the MAC peripheral for the duration of SMI accesses
